# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Online market IPi lari",
    "version": "1.0.0",
    "description": "Bu shaxsiy online market web siti uchun qilingan loyiha IPI lari"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/products/all": {
      "get": {
        "tags": ["Products"],
        "summary": "Hamma mahsulotlarni olish uchun",
        "description": "Hamma mahsulot propertylarini olish uchun, Pagenation qilish va category boyicha filter qo'yilgan bu kattaliklar query params orqali beriladi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The page number",
            "type": "integer",
            "required": false,
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The number of items per page",
            "type": "integer",
            "required": false,
            "default": 10
          },
          {
            "name": "category",
            "in": "query",
            "description": "The category of items to retrieve",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Bitta mahsulotni olish uchun API",
        "description": "Kiritilgan ID boyicha ma'lumot olishingiz mumin",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/": {
      "post": {
        "tags": ["Products"],
        "summary": "Maxsulot qoshish uchun",
        "description": "Maxsulot qo'shayotgan foydalanuvchi seller rolida bo'lishi talab qilinadi",
        "produces": ["multipart/form-data"],
        "parameters": [
          {
            "name": "image",
            "in": "form-data",
            "required": true,

            "type": "file"
          },
          {
            "name": "product",
            "in": "body",
            "required": true,
            "type": "object",
            "default": {
              "title": "title",
              "description": "description",
              "price": 5000,
              "category": "category",
              "image": "file"
            }
          },

          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/delete/{id}": {
      "delete": {
        "tags": ["Products"],
        "summary": "Sotuvchilar O'zining qo'shgan maxsulotlari o'chirib yuborishi mumkin",
        "description": "Sotuvchi roliga ega bo'lishi talab qilinadi, xabar emailga jo'natiladi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/products/update/{id}": {
      "put": {
        "tags": ["Products"],
        "summary": "Sotuvchilar O'zining qo'shgan maxsulotlari o'zgartirishlari mumkin",
        "description": "Sotuvchi roliga ega bo'lishi talab qilinadi, xabar emailga jo'natiladi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Bu o'zgartirishni yuborish",
            "schema": {
              "type": "object",
              "default": {
                "title": "title",
                "description": "description",
                "price": 2000,
                "category": "category"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "default": [
                    {
                      "product": "Product object"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{id}": {
      "post": {
        "tags": ["Orders"],
        "summary": "Yangi buyurtma qo'shish uchun",
        "description": "Har bir mahsulot tanlanganda ushbu API ga so'rov jo'natiladi",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "quantity",
            "in": "body",
            "required": true,
            "type": "number",
            "default": {
              "quantity": 1
            }
          }
        ],
        "responses": {
          "201": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "default": {
                    "user": {
                      "email": "email",
                      "username": "username"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/my-basket": {
      "get": {
        "tags": ["Orders"],
        "summary": "Savatni olish uchun",
        "description": "Mening buyurtmalarim",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "A successful response",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "default": {
                    "user": {
                      "email": "email",
                      "username": "username"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}